Modularize common.rs structs
